Basic movement
The player moves on the vertical, horizontal and both diagonal axises. The diagonal speed is calculated to not be faster than on the horizontal and vertical axis. The diagonal speed is rounded up after the calculation, so it can be a bit faster than horizontal or vertical, but this is necesary, since pygame cant move object by fractions of pixels.