Map choice, pause and trz again
The players now get to choose the map they want to play on. The choice is made with the cursor and confirmed with LMB. Players can pause the game during the match with the P key and unpause with the ESC key. At the end of the game, the players can choose to play again, whiping all the sprites, creating them again and going back to the map choice.